+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Optimized `EventStateMachineFactory` by no longer using callback factory abstraction.
+- `EventStateMachine` now receives and uses an event factory to create transition events.
+
+### Fixed
+- Static event params given to `EventStateMachine` no longer override transition-related event params.
 
 ## [0.1-alpha1] - 2018-05-15
 Initial version.